@@ -21,7 +21,7 @@
     "requireindex": "1.2.0"
   },
   "devDependencies": {
-    "eslint": "~7.6.0",
+    "eslint": "~9.0.0",
     "mocha": "~8.1.1"
   },
   "engines": {
