@@ -121,7 +121,7 @@ module.exports = {
                     if (callee.async || getTypeAnnotation(callee.returnType) === "Promise") {
                         context.report({
                             fix(fixer) {
-                                return fixer.replaceText(node, `await ${context.getSourceCode().getText(node)}`);
+                                return fixer.replaceText(node, `await ${context.sourceCode.getText(node)}`);
                             },
                             loc: node.loc,
                             messageId: "foundFloating",
@@ -145,7 +145,7 @@ module.exports = {
 
                 // handle named function calls
                 if (callee.type === "Identifier") {
-                    const variable = findInScope(context.getScope(node), callee.name);
+                    const variable = findInScope(context.sourceCode.getScope(node), callee.name);
                     if (variable === null || typeof variable === "undefined") {
                         return;
                     }
@@ -153,7 +153,7 @@ module.exports = {
                     if (variable.async || getTypeAnnotation(variable) === "Promise") {
                         context.report({
                             fix(fixer) {
-                                return fixer.replaceText(node, `await ${context.getSourceCode().getText(node)}`);
+                                return fixer.replaceText(node, `await ${context.sourceCode.getText(node)}`);
                             },
                             loc: node.loc,
                             messageId: "foundFloating",
