@@ -18,20 +18,20 @@ const path = require("path"),
 //------------------------------------------------------------------------------
 
 /**
- * Gets the path to the specified parser.
+ * Gets the specified parser.
  *
  * @param {string} name - The parser name to get.
- * @returns {string} The path to the specified parser.
+ * @returns {function} specificed parser
  */
 function parser(name) {
-    return path.resolve(__dirname, `../../fixtures/parsers/no-floating-promise/${name}.js`);
+    return require(`../../fixtures/parsers/no-floating-promise/${name}.js`);
 }
 
 //------------------------------------------------------------------------------
 // Tests
 //------------------------------------------------------------------------------
 
-const ruleTester = new RuleTester({ parserOptions: { ecmaVersion: 8, sourceType: "module" } });
+const ruleTester = new RuleTester({ languageOptions: { ecmaVersion: 8, sourceType: "module" } });
 
 ruleTester.run("no-floating-promise", rule, {
     valid: [
@@ -39,44 +39,46 @@ ruleTester.run("no-floating-promise", rule, {
 
             // awaiting a function identifier works
             code: "async function foo() {}; async function wrap() { await foo(); }",
-            parserOptions: { ecmaVersion: 8 }
+            languageOptions: { ecmaVersion: 8 }
         },
         {
 
             // if not awaited but used in an assignment, no error is reported
             code: "async function foo() {}; async function wrap() { const a = foo(); }",
-            parserOptions: { ecmaVersion: 8 }
+            languageOptions: { ecmaVersion: 8 }
         },
         {
 
             // if not awaited but returned, no error is reported
             code: "async function foo() {}; function bar() { return foo(); };",
-            parserOptions: { ecmaVersion: 8 }
+            languageOptions: { ecmaVersion: 8 }
         },
         {
 
             // awaiting a non-async function that returns a Promise works
             code: "function foo(): Promise<void> { return Promise.resolve(); }; async function wrap() { await foo(); }",
-            parser: parser("no-floating-promise-typed"),
-            parserOptions: { ecmaVersion: 8 }
+            languageOptions: {
+              ecmaVersion: 8,
+              parser: parser("no-floating-promise-typed")
+            }
         },
         {
 
             // awaiting a FunctionExpression works
             code: "async function wrap() { await (async function () { return 5; })(); };",
-            parserOptions: { ecmaVersion: 8 }
+            languageOptions: { ecmaVersion: 8 }
         },
         {
 
             // awaiting an ArrowFunctionExpression works
             code: "async function wrap() { await (async () => 5)(); };",
-            parserOptions: { ecmaVersion: 8 }
+            languageOptions: { ecmaVersion: 8 }
         },
         {
 
             // doesn't break regular function calls
             code: "function foo() {}; foo();",
-            parserOptions: { ecmaVersion: 8 }
+            languageOptions: { ecmaVersion: 8 }
         },
         {
 
@@ -85,20 +87,22 @@ ruleTester.run("no-floating-promise", rule, {
              * means unfortunately we won"t catch it
              */
             code: "function foo() { return Promise.resolve(); }; async function wrap() { foo(); }",
-            parserOptions: { ecmaVersion: 8 }
+            languageOptions: { ecmaVersion: 8 }
         },
         {
 
             // doesn't crash when type "any" is used
             code: "function foo(): any { return Promise.resolve(); }; async function wrap() { foo(); }",
-            parser: parser("floating-promise-any"),
-            parserOptions: { ecmaVersion: 8 }
+            languageOptions: {
+              ecmaVersion: 8,
+              parser: parser("floating-promise-any")
+            }
         },
         {
 
             // no error if then is used
             code: "async function foo() {}; async function wrap() { foo().then(() => {}); }",
-            parserOptions: { ecmaVersion: 8 }
+            languageOptions: { ecmaVersion: 8 }
         },
         {
 
@@ -107,8 +111,10 @@ ruleTester.run("no-floating-promise", rule, {
         },
         {
             code: 'const foo = async ( lambda: void=>void ) { lambda(); }',
-            parser: parser("no-floating-promise-argument"),
-            parserOptions: { ecmaVersion: 8 }
+            languageOptions: {
+              ecmaVersion: 8,
+              parser: parser("no-floating-promise-argument")
+            }
         }
     ],
     invalid: [
@@ -117,7 +123,7 @@ ruleTester.run("no-floating-promise", rule, {
             // fails if missing an await on a regular function call
             code: "async function foo() {}; async function wrap() { foo(); }",
             output: "async function foo() {}; async function wrap() { await foo(); }",
-            parserOptions: { ecmaVersion: 8 },
+            languageOptions: { ecmaVersion: 8 },
             errors: [{ messageId: "foundFloating" }]
         },
         {
@@ -125,7 +131,7 @@ ruleTester.run("no-floating-promise", rule, {
             // fails if missing an await on a regular function call wrapped in something
             code: "async function foo() {}; async function wrap() { (foo)(); }",
             output: "async function foo() {}; async function wrap() { await (foo)(); }",
-            parserOptions: { ecmaVersion: 8 },
+            languageOptions: { ecmaVersion: 8 },
             errors: [{ messageId: "foundFloating" }]
         },
         {
@@ -133,8 +139,10 @@ ruleTester.run("no-floating-promise", rule, {
             // fails if you"re missing an await on a non-async function that returns a Promise
             code: "function foo(): Promise<void> { return Promise.resolve(); }; async function wrap() { foo(); }",
             output: "function foo(): Promise<void> { return Promise.resolve(); }; async function wrap() { await foo(); }",
-            parser: parser("floating-promise-typed-identifier"),
-            parserOptions: { ecmaVersion: 8 },
+            languageOptions: {
+              ecmaVersion: 8,
+              parser: parser("floating-promise-typed-identifier")
+            },
             errors: [{ messageId: "foundFloating" }]
         },
         {
@@ -142,7 +150,7 @@ ruleTester.run("no-floating-promise", rule, {
             // fails when missing an await on a FunctionExpression
             code: "async function wrap() { (async function () { return 5; })(); };",
             output: "async function wrap() { await (async function () { return 5; })(); };",
-            parserOptions: { ecmaVersion: 8 },
+            languageOptions: { ecmaVersion: 8 },
             errors: [{ messageId: "foundFloating" }]
         },
         {
@@ -150,7 +158,7 @@ ruleTester.run("no-floating-promise", rule, {
             // fails when missing an await on an ArrowFunctionExpression
             code: "async function wrap() { (async () => 5)(); };",
             output: "async function wrap() { await (async () => 5)(); };",
-            parserOptions: { ecmaVersion: 8 },
+            languageOptions: { ecmaVersion: 8 },
             errors: [{ messageId: "foundFloating" }]
         },
         {
@@ -158,8 +166,10 @@ ruleTester.run("no-floating-promise", rule, {
             // fails when missing an await on a non-async FunctionExpression that returns a Promise
             code: "async function wrap() { (function (): Promise<void> { return Promise.resolve(); })(); };",
             output: "async function wrap() { await (function (): Promise<void> { return Promise.resolve(); })(); };",
-            parser: parser("floating-promise-typed-expression"),
-            parserOptions: { ecmaVersion: 8 },
+            languageOptions: {
+              ecmaVersion: 8,
+              parser: parser("floating-promise-typed-expression")
+            },
             errors: [{ messageId: "foundFloating" }]
         },
         {
@@ -167,8 +177,10 @@ ruleTester.run("no-floating-promise", rule, {
             // calling a lambda function argument
             code: 'class Foo { foo = async ( lambda: (void=>Promise<void>) ) => { lambda(); }}',
             output: 'class Foo { foo = async ( lambda: (void=>Promise<void>) ) => { await lambda(); }}',
-            parser: parser("floating-promise-argument"),
-            parserOptions: { ecmaVersion: 8 },
+            languageOptions: {
+              ecmaVersion: 8,
+              parser: parser("floating-promise-argument")
+            },
             errors: [{ messageId: "foundFloating" }]
         },
     ]
